@@ -26,22 +26,14 @@ final class UnnecessaryFunctionCallValidator implements AfterEveryFunctionCallAn
         'printf',
     ];
 
-    /** @psalm-var non-empty-string */
-    private string $functionName;
-
-    /** @psalm-var non-empty-list<Arg> */
-    private array $arguments;
-
     /**
      * @param non-empty-string    $functionName
      * @param non-empty-list<Arg> $arguments
      */
     public function __construct(
-        string $functionName,
-        array $arguments
+        private string $functionName,
+        private array $arguments,
     ) {
-        $this->functionName = $functionName;
-        $this->arguments    = $arguments;
     }
 
     public static function afterEveryFunctionCallAnalysis(AfterEveryFunctionCallAnalysisEvent $event): void
@@ -83,7 +75,7 @@ final class UnnecessaryFunctionCallValidator implements AfterEveryFunctionCallAn
         StatementsSource $statementsSource,
         CodeLocation $codeLocation,
         Context $context,
-        int $phpVersion
+        int $phpVersion,
     ): void {
         $template = $this->arguments[0];
 
@@ -96,7 +88,7 @@ final class UnnecessaryFunctionCallValidator implements AfterEveryFunctionCallAn
                 false,
                 $statementsSource,
             );
-        } catch (InvalidArgumentException $exception) {
+        } catch (InvalidArgumentException) {
             return;
         }
 
@@ -108,7 +100,7 @@ final class UnnecessaryFunctionCallValidator implements AfterEveryFunctionCallAn
 
     private function assertFunctionCallMakesSense(
         CodeLocation $codeLocation,
-        TemplatedStringParser $parsed
+        TemplatedStringParser $parsed,
     ): void {
         if ($parsed->getTemplate() !== $parsed->getTemplateWithoutPlaceholder()) {
             return;
