@@ -9,15 +9,11 @@ use Psalm\Issue\PluginIssue;
 
 final class UnnecessaryFunctionCall extends PluginIssue
 {
-    public string $function_id;
-
-    public function __construct(CodeLocation $code_location, string $function_id)
+    public function __construct(CodeLocation $code_location, public string $function_id)
     {
         parent::__construct(
             'Function call is unnecessary as there is no placeholder within the template.',
             $code_location,
         );
-
-        $this->function_id = $function_id;
     }
 }
